[1.2][DC-385] Changing Geographical behavior to use double type instead of float
git-svn-id: file:///var/www/production/svnmirror/svn_doctrine/branches/1.2/lib@7253 2b2ae520-0f61-44a6-8c59-7e116ffd6265

+++ Doctrine/Template/Geographical.php
@@ -40,11 +40,11 @@ class Doctrine_Template_Geographical extends Doctrine_Template
      * @var string
      */
     protected $_options = array('latitude' =>  array('name'     =>  'latitude',
-                                                     'type'     =>  'float',
+                                                     'type'     =>  'double',
                                                      'size'     =>  null,
                                                      'options'  =>  array()),
                                 'longitude' => array('name'     =>  'longitude',
-                                                     'type'     =>  'float',
+                                                     'type'     =>  'double',
                                                      'size'     =>  null,
                                                      'options'  =>  array()));
